rename folder to item-variant
