+++ packages/pwa/app/components/item-variant/item-attributes.jsx
