+++ packages/pwa/app/components/item-variant/item-image.jsx
