+++ packages/pwa/app/components/item-variant/item-name.jsx
