+++ packages/pwa/app/components/order-summary/index.jsx
@@ -12,11 +12,11 @@ import useBasket from '../../commerce-api/hooks/useBasket'
 import {BasketIcon, ChevronDownIcon, ChevronUpIcon} from '../icons'
 import Link from '../link'
 import {PromoCode, usePromoCode} from '../promo-code'
-import ItemVariantProvider from '../cart-item-variant'
-import CartItemVariantImage from '../cart-item-variant/item-image'
-import CartItemVariantName from '../cart-item-variant/item-name'
-import CartItemVariantAttributes from '../cart-item-variant/item-attributes'
-import CartItemVariantPrice from '../cart-item-variant/item-price'
+import ItemVariantProvider from '../item-variant'
+import CartItemVariantImage from '../item-variant/item-image'
+import CartItemVariantName from '../item-variant/item-name'
+import CartItemVariantAttributes from '../item-variant/item-attributes'
+import CartItemVariantPrice from '../item-variant/item-price'
 import PromoPopover from '../promo-popover'
 
 const CartItems = ({basket}) => {
